code cleanup: favor braces when blocks have mixed brace use.

+++ source/blender/editors/interface/interface_style.c
@@ -216,7 +216,7 @@ void uiStyleFontDrawRotated(uiFontStyle *fs, rcti *rect, const char *str)
 	/* rotate counter-clockwise for now (assumes left-to-right language)*/
 	xofs += height;
 	yofs = BLF_width(fs->uifont_id, str) + 5;
-	angle = M_PI / 2.0f;
+	angle = (float)M_PI / 2.0f;
 
 	/* translate rect to vertical */
 	txtrect.xmin = rect->xmin - BLI_rcti_size_y(rect);
